pyproject.toml
[project]
name = "parsec"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.0",
    "groq>=1.0.0",
    "httpx==0.27.2",
    "itsdangerous==2.2.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "python-jose[cryptography]==3.3.0",
    "uvicorn[standard]==0.30.6",
]
